separated tests into own project
